Expose CSRF Token refresh as a function
Refreshing CSRF tokens ensures that forms do not contain cached tokens.
This is currently done once, when jquery-ujs is loaded. If a cached form
is later loaded onto the page (say via a remote form for 'load more'),
the CSRF token is not udated.

This change exposes the CSRF token refreshing code at
`$.rails.refreshCSRFTokens` so it can be called at will without developers
having to re-implement the function.