data-method whitelist to block adding CSRF
If an attacker has the ability to create link tags and attributes they
have the ability to intercept the CSRF token, even if CSP is enabled.
By setting the href to another site and adding `data-method='post'` the
CSRF token will automatically be added to the request and sent to an
attackers site. It is important to note that this is only applicable to
`handleMethod` (`data-method`) call as when done through `handleRemote`
(`data-remote`) the CSP will block the request to the attackers domain.

This fix adds a configurable whitelist of hrefs to verify the link
against before adding the CSRF token. The default is to allow
everything to avoid break existing clients.