Bug where one radio button in group required
The HTML spec does not seemd to clearly state what "required" does when
only some of a radio button set are required.

This answer on StackOverflow states that only one of the set is
required: http://stackoverflow.com/a/8287947/1009332

Per the comments added in the fix:

We cannot just compare allInputs for radio buttons, as the spec for required radio buttons
states that only one of a radio button set needs to be required.

If you have 2 radio buttons, and one is checked, but it's not required, then we get
the other required radio button and see that there are no other radio buttons checked
for the "required" ones.

Thus, if the input control is a radio button, then filter for all radio buttons with
the same name on the form.

This will get all radio buttons of a given name that are checked and see that at
least one is checked. Don't count unchecked required radio if other radio with same
name is checked. But do count checked non-required radio of the same name.

Note, each radio button name that is required is checked only once.